requirements.txt
aiogram
google-genai
requests